+++ src/mumax/engine/solve_euler_backward.go
@@ -2,7 +2,7 @@
 //  Copyright 2011  Arne Vansteenkiste and Ben Van de Wiele.
 //  Use of this source code is governed by the GNU General Public License version 3
 //  (as published by the Free Software Foundation) that can be found in the license.txt file.
-//  Note that you are welcome to modify this code under the condition that you do not remove any 
+//  Note that you are welcome to modify this code under the condition that you do not remove any
 //  copyright notices and prominently state that you modified it, giving a relevant date.
 
 package engine
@@ -19,7 +19,7 @@ import (
 type BDFEuler struct {
 	ybuffer    []*gpu.Array // initial derivative
 	y0buffer   []*gpu.Array // initial derivative
-	err 	   []*Quant     // error estimates for each equation
+	err        []*Quant     // error estimates for each equation
 	maxIterErr []*Quant     // error estimates for each equation
 	maxIter    []*Quant     // maximum number of iterations per step
 	diff       []gpu.Reductor
@@ -37,57 +37,56 @@ func (s *BDFEuler) Step() {
 
 	// get dt here to avoid updates later on.
 	dt := engine.dt.Scalar()
-	// Advance time and update all inputs  
+	// Advance time and update all inputs
 	e.time.SetScalar(e.time.Scalar() + dt)
 
 	// Then step all outputs (without intermediate updates!)
 	// and invalidate them.
-	
 
 	for i := range equation {
 		err := 1.0e38
 		s.iterations.SetScalar(0)
 		iter := 0
-		
-		// Do forward Euler step 
+
+		// Do forward Euler step
 		// Zero order approximation
 		y := equation[i].output[0]
 		dy := equation[i].input[0]
 		dyMul := dy.multiplier
-		h := float32(dt * dyMul[0])
+		h := dt * dyMul[0]
 		s.y0buffer[i].CopyFromDevice(y.Array()) // save for later
 		gpu.Madd(y.Array(), s.y0buffer[i], dy.Array(), h)
 		y.Invalidate()
 		equation[i].input[0].Update()
 		s.iterations.SetScalar(s.iterations.Scalar() + 1)
-		
+
 		// Do backward Euler step and solve it
 		// Do higher order approximation until converges
 		// Using fixed-point iterator
-		
+
 		maxIterErr := s.maxIterErr[i].Scalar()
 		maxIter := int(s.maxIter[i].Scalar())
-			
+
 		for err > maxIterErr {
-				gpu.Madd(s.ybuffer[i], s.y0buffer[i], dy.Array(), h) 
-				err = float64(s.diff[i].MaxDiff(y.Array(), s.ybuffer[i]))
-				sum := float64(s.diff[i].MaxSum(y.Array(), s.ybuffer[i]))
-				if sum > 0.0 {
-					err = err / sum
-				}
-				iter = iter + 1
-				s.iterations.SetScalar(s.iterations.Scalar() + 1)
-				y.Array().CopyFromDevice(s.ybuffer[i])
-				y.Invalidate()
-				equation[i].input[0].Update()
-				if iter > maxIter {
-					panic(Bug(fmt.Sprintf("The BDF iterator cannot converge for %s! Please decrease the time step and re-run!",y.Name())))
-				}
+			gpu.Madd(s.ybuffer[i], s.y0buffer[i], dy.Array(), h)
+			err = float64(s.diff[i].MaxDiff(y.Array(), s.ybuffer[i]))
+			sum := float64(s.diff[i].MaxSum(y.Array(), s.ybuffer[i]))
+			if sum > 0.0 {
+				err = err / sum
+			}
+			iter = iter + 1
+			s.iterations.SetScalar(s.iterations.Scalar() + 1)
+			y.Array().CopyFromDevice(s.ybuffer[i])
+			y.Invalidate()
+			equation[i].input[0].Update()
+			if iter > maxIter {
+				panic(Bug(fmt.Sprintf("The BDF iterator cannot converge for %s! Please decrease the time step and re-run!", y.Name())))
+			}
 		}
-		
+
 	}
 
-	// Advance step	
+	// Advance step
 	e.step.SetScalar(e.step.Scalar() + 1) // advance step
 }
 
@@ -118,7 +117,7 @@ func LoadBDFEuler(e *Engine) {
 	s.diff = make([]gpu.Reductor, len(equation))
 	s.maxIter = make([]*Quant, len(equation))
 	s.maxIterErr = make([]*Quant, len(equation))
-	
+
 	for i := range equation {
 
 		eqn := &(equation[i])
@@ -128,15 +127,14 @@ func LoadBDFEuler(e *Engine) {
 		s.err[i] = e.AddNewQuant(out.Name()+"_error", SCALAR, VALUE, unit, "Error/step estimate for "+out.Name())
 		s.maxIter[i] = e.AddNewQuant(out.Name()+"_maxIterations", SCALAR, VALUE, unit, "Maximum number of evaluations per step"+out.Name())
 		s.maxIterErr[i] = e.AddNewQuant(out.Name()+"_maxIterError", SCALAR, VALUE, unit, "The maximum error of iterator"+out.Name())
-		
+
 		s.maxIterErr[i].SetScalar(1e-5)
 		s.maxIter[i].SetScalar(3)
-		
+
 		s.diff[i].Init(out.Array().NComp(), out.Array().Size3D())
 		s.maxIter[i].SetVerifier(Positive)
 		s.maxIterErr[i].SetVerifier(Positive)
-		
-		
+
 		// TODO: recycle?
 
 		y := equation[i].output[0]
